Refactor + UI fixes + map improvements

+++ app/src/main/java/io/github/azakidev/move/data/db/entities/StopEntity.kt
@@ -26,8 +26,8 @@ data class StopEntity(
 
     val lines: List<Int>,
 
-    val geoX: Float?,
-    val geoY: Float?,
+    val geoX: Double?,
+    val geoY: Double?,
 
     val notifications: List<String>,
 )
